PythonUpsplusReader/requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.9
mock==4.0.3
pi-ina219==1.4.0
pip==22.1.1
setuptools==62.3.2
smbus2==0.4.1
wheel==0.37.1
